Add internal function that fails on missing config files
poptReadConfigFile reads an optional config file, ie. it returns success
even if that config file does not exist. This behavior is exported and
can not be changed.

But sometimes, eg. with ~/.config/popt and ~/.popt, we need to know
whether a config file exists in order to read just one. Add
poptReadConfigFileInternal to allow that. It basically is the guts of
the old poptReadConfigFile. And make poptReadConfigFile a very thin
wrapper around poptReadConfigFileInternal.